+++ src/Tickets/NotEvenlySplitDecorator.java
@@ -7,18 +7,18 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class NotEvenlySplitDecorator implements TicketDecorator{
-    private Map<Person,Float> detailedPaidFor;
+    private Map<Person,Double> detailedPaidFor;
     private Person person;
-    public NotEvenlySplitDecorator(Person person, Map<Person,Float> detailedPaidFor) {
+    public NotEvenlySplitDecorator(Person person, Map<Person,Double> detailedPaidFor) {
         this.detailedPaidFor = detailedPaidFor;
         this.person = person;
     }
 
     @Override
-    public Map<Person,Float> getBalanceDifferencePerPerson() {
-        Map<Person,Float> differencePerPerson = new HashMap<Person, Float>();
-        float totalPrice = 0;
-        for (Map.Entry<Person, Float> entry : detailedPaidFor.entrySet()) {
+    public Map<Person,Double> getBalanceDifferencePerPerson() {
+        Map<Person,Double> differencePerPerson = new HashMap<Person, Double>();
+        double totalPrice = 0;
+        for (Map.Entry<Person, Double> entry : detailedPaidFor.entrySet()) {
             if (!(entry.getKey() == person)) {
                 differencePerPerson.put(entry.getKey(), - entry.getValue());
                 totalPrice += entry.getValue();
