+++ src/Tickets/RestaurantTicket.java
@@ -5,7 +5,7 @@ import Person.Person;
 import java.util.ArrayList;
 
 public class RestaurantTicket extends Ticket{
-    public RestaurantTicket(float price, TicketDecorator ticketDecorator, Person person) {
+    public RestaurantTicket(double price, TicketDecorator ticketDecorator, Person person) {
         super(price, ticketDecorator, person);
     }
 }
