+++ src/Tickets/TaxiTicket.java
@@ -5,7 +5,7 @@ import Person.Person;
 import java.util.ArrayList;
 
 public class TaxiTicket extends Ticket{
-    public TaxiTicket(float price, TicketDecorator ticketDecorator, Person person) {
+    public TaxiTicket(double price, TicketDecorator ticketDecorator, Person person) {
         super(price, ticketDecorator, person);
     }
 }
